去除新添加的api
