+++ common-api/src/main/java/io/seata/commonapi/annotation/BusinessActionContextParameterDesc.java
@@ -16,6 +16,7 @@
 package io.seata.commonapi.annotation;
 
 import io.seata.common.util.StringUtils;
+import io.seata.rm.tcc.api.BusinessActionContextParameter;
 
 import java.lang.annotation.Annotation;
 
@@ -39,13 +40,6 @@ public class BusinessActionContextParameterDesc {
         if (annotation instanceof BusinessActionContextParameter) {
             businessActionContextParameterDesc = new BusinessActionContextParameterDesc();
             BusinessActionContextParameter businessActionContextParameter = (BusinessActionContextParameter) annotation;
-            businessActionContextParameterDesc.setIndex(businessActionContextParameter.index());
-            businessActionContextParameterDesc.setParamInProperty(businessActionContextParameter.isParamInProperty());
-            businessActionContextParameterDesc.setParamName(StringUtils.isNotBlank(businessActionContextParameter.paramName()) ? businessActionContextParameter.paramName() : businessActionContextParameter.value());
-
-        } else if (annotation instanceof io.seata.rm.tcc.api.BusinessActionContextParameter) {
-            businessActionContextParameterDesc = new BusinessActionContextParameterDesc();
-            io.seata.rm.tcc.api.BusinessActionContextParameter businessActionContextParameter = (io.seata.rm.tcc.api.BusinessActionContextParameter) annotation;
             businessActionContextParameterDesc.setIndex(businessActionContextParameter.index());
             businessActionContextParameterDesc.setParamInProperty(businessActionContextParameter.isParamInProperty());
             businessActionContextParameterDesc.setParamName(StringUtils.isNotBlank(businessActionContextParameter.paramName()) ? businessActionContextParameter.paramName() : businessActionContextParameter.value());
