+++ common-api/src/main/java/io/seata/commonapi/interceptor/ActionContextUtil.java
@@ -20,9 +20,8 @@ import io.seata.common.exception.FrameworkException;
 import io.seata.common.util.CollectionUtils;
 import io.seata.common.util.ReflectionUtil;
 import io.seata.common.util.StringUtils;
-import io.seata.commonapi.annotation.BusinessActionContextParameter;
-import io.seata.commonapi.annotation.BusinessActionContextParameterDesc;
-import io.seata.commonapi.api.BusinessActionContext;
+import io.seata.rm.tcc.api.BusinessActionContextParameter;
+import io.seata.rm.tcc.api.BusinessActionContext;
 import io.seata.rm.tcc.api.ParamType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -70,11 +69,8 @@ public final class ActionContextUtil {
             Map<String, Object> context = new HashMap<>(8);
             for (Field f : fields) {
                 // get annotation
-                BusinessActionContextParameterDesc businessActionContextParameterDesc = BusinessActionContextParameterDesc.createFromBusinessActionContextParameter(f.getAnnotation(BusinessActionContextParameter.class));
-                if (businessActionContextParameterDesc == null) {
-                    businessActionContextParameterDesc = BusinessActionContextParameterDesc.createFromBusinessActionContextParameter(f.getAnnotation(io.seata.rm.tcc.api.BusinessActionContextParameter .class));
-                }
-                if (businessActionContextParameterDesc == null) {
+                BusinessActionContextParameter annotation = f.getAnnotation(BusinessActionContextParameter.class);
+                if (annotation == null) {
                     continue;
                 }
 
@@ -84,7 +80,7 @@ public final class ActionContextUtil {
 
                 // load param by the config of annotation, and then put into the context
                 String fieldName = f.getName();
-                loadParamByAnnotationAndPutToContext(ParamType.FIELD, fieldName, fieldValue, businessActionContextParameterDesc, context);
+                loadParamByAnnotationAndPutToContext(ParamType.FIELD, fieldName, fieldValue, annotation, context);
             }
             return context;
         } catch (Exception e) {
@@ -98,17 +94,17 @@ public final class ActionContextUtil {
      * @param paramType     the param type, 'param' or 'field'
      * @param paramName     the param name
      * @param paramValue    the param value
-     * @param businessActionContextParameterDesc    the annotation on the param or field
+     * @param annotation    the annotation on the param or field
      * @param actionContext the action context
      */
     public static void loadParamByAnnotationAndPutToContext(@Nonnull final ParamType paramType, @Nonnull String paramName, Object paramValue,
-                                                            @Nonnull final BusinessActionContextParameterDesc businessActionContextParameterDesc, @Nonnull final Map<String, Object> actionContext) {
+            @Nonnull final BusinessActionContextParameter annotation, @Nonnull final Map<String, Object> actionContext) {
         if (paramValue == null) {
             return;
         }
 
         // If {@code index >= 0}, get by index from the list param or field
-        int index = businessActionContextParameterDesc.getIndex();
+        int index = annotation.index();
         if (index >= 0) {
             paramValue = getByIndex(paramType, paramName, paramValue, index);
             if (paramValue == null) {
@@ -117,14 +113,14 @@ public final class ActionContextUtil {
         }
 
         // if {@code isParamInProperty == true}, fetch context from paramValue
-        if (businessActionContextParameterDesc.isParamInProperty()) {
+        if (annotation.isParamInProperty()) {
             Map<String, Object> paramContext = fetchContextFromObject(paramValue);
             if (CollectionUtils.isNotEmpty(paramContext)) {
                 actionContext.putAll(paramContext);
             }
         } else {
             // get param name from the annotation
-            String paramNameFromAnnotation = businessActionContextParameterDesc.getParamName();
+            String paramNameFromAnnotation = getParamNameFromAnnotation(annotation);
             if (StringUtils.isNotBlank(paramNameFromAnnotation)) {
                 paramName = paramNameFromAnnotation;
             }
@@ -157,14 +153,6 @@ public final class ActionContextUtil {
     }
 
     public static String getParamNameFromAnnotation(@Nonnull BusinessActionContextParameter annotation) {
-        String paramName = annotation.paramName();
-        if (StringUtils.isBlank(paramName)) {
-            paramName = annotation.value();
-        }
-        return paramName;
-    }
-
-    public static String getParamNameFromAnnotation(@Nonnull io.seata.rm.tcc.api.BusinessActionContextParameter annotation) {
         String paramName = annotation.paramName();
         if (StringUtils.isBlank(paramName)) {
             paramName = annotation.value();
