+++ common-api/src/main/java/io/seata/commonapi/interceptor/parser/GlobalTransactionalInterceptorParser.java
@@ -16,16 +16,15 @@
 package io.seata.commonapi.interceptor.parser;
 
 import io.seata.common.util.CollectionUtils;
-import io.seata.commonapi.annotation.GlobalLock;
-import io.seata.commonapi.annotation.GlobalTransactional;
 import io.seata.commonapi.interceptor.handler.GlobalTransactionalInterceptorHandler;
 import io.seata.commonapi.interceptor.handler.ProxyInvocationHandler;
 import io.seata.config.ConfigurationCache;
 import io.seata.config.ConfigurationChangeListener;
 import io.seata.core.constants.ConfigurationKeys;
+import io.seata.spring.annotation.GlobalLock;
+import io.seata.spring.annotation.GlobalTransactional;
 import io.seata.tm.api.DefaultFailureHandlerImpl;
 
-import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.util.HashSet;
 import java.util.Set;
@@ -58,6 +57,7 @@ public class GlobalTransactionalInterceptorParser implements InterfaceParser {
             ConfigurationCache.addConfigListener(ConfigurationKeys.DISABLE_GLOBAL_TRANSACTION, (ConfigurationChangeListener) proxyInvocationHandler);
             return proxyInvocationHandler;
         }
+
         return null;
     }
 
@@ -68,11 +68,7 @@ public class GlobalTransactionalInterceptorParser implements InterfaceParser {
                 if (clazz == null) {
                     continue;
                 }
-                Annotation trxAnno = clazz.getAnnotation(GlobalTransactional.class);
-                if (trxAnno != null) {
-                    return true;
-                }
-                trxAnno = clazz.getAnnotation(io.seata.spring.annotation.GlobalTransactional.class);
+                GlobalTransactional trxAnno = clazz.getAnnotation(GlobalTransactional.class);
                 if (trxAnno != null) {
                     return true;
                 }
@@ -82,23 +78,12 @@ public class GlobalTransactionalInterceptorParser implements InterfaceParser {
                     if (trxAnno != null) {
                         methodsToProxy.add(method.getName());
                         result = true;
-                    } else {
-                        trxAnno = method.getAnnotation(io.seata.spring.annotation.GlobalTransactional.class);
-                        if (trxAnno != null) {
-                            methodsToProxy.add(method.getName());
-                            result = true;
-                        }
                     }
-                    Annotation lockAnno = method.getAnnotation(GlobalLock.class);
+
+                    GlobalLock lockAnno = method.getAnnotation(GlobalLock.class);
                     if (lockAnno != null) {
                         methodsToProxy.add(method.getName());
                         result = true;
-                    } else {
-                        lockAnno = method.getAnnotation(io.seata.spring.annotation.GlobalLock.class);
-                        if (lockAnno != null) {
-                            methodsToProxy.add(method.getName());
-                            result = true;
-                        }
                     }
                 }
             }
