+++ tcc/src/test/java/io/seata/rm/tcc/NormalTccActionImpl.java
@@ -15,7 +15,8 @@
  */
 package io.seata.rm.tcc;
 
-import io.seata.commonapi.api.BusinessActionContext;
+
+import io.seata.rm.tcc.api.BusinessActionContext;
 
 import java.util.List;
