+++ tcc/src/test/java/io/seata/rm/tcc/interceptor/ActionInterceptorHandlerTest.java
@@ -18,7 +18,7 @@ package io.seata.rm.tcc.interceptor;
 import io.seata.commonapi.interceptor.ActionInterceptorHandler;
 import io.seata.rm.tcc.TccAction;
 import io.seata.rm.tcc.TccParam;
-import io.seata.commonapi.api.BusinessActionContext;
+import io.seata.rm.tcc.api.BusinessActionContext;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
