+++ tcc/src/test/java/io/seata/rm/tcc/interceptor/parser/TccActionInterceptorParserTest.java
@@ -17,8 +17,6 @@ package io.seata.rm.tcc.interceptor.parser;
 
 import io.seata.commonapi.interceptor.handler.ProxyInvocationHandler;
 import io.seata.rm.tcc.NormalTccActionImpl;
-import io.seata.rm.tcc.TccAction;
-import io.seata.rm.tcc.TccActionImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
