+++ tcc/src/test/java/io/seata/rm/tcc/spring/tcc/TccAction.java
@@ -15,7 +15,7 @@
  */
 package io.seata.rm.tcc.spring.tcc;
 
-import io.seata.commonapi.api.BusinessActionContext;
+import io.seata.rm.tcc.api.BusinessActionContext;
 import io.seata.rm.tcc.api.TwoPhaseBusinessAction;
 
 /**
